Extract user invite feature into it's own replaceable components
