+++ plugins/UsersManager/stylesheets/usersManager.less
@@ -45,7 +45,7 @@
   .card h2.card-title, .card h2 .card-title {
     padding-left: 15px !important;
   }
-  .userEditForm {
+  .userEditForm, .userInviteForm {
     .card-title {
       margin-top: 45px;
     }
