+++ plugins/UsersManager/templates/index.twig
@@ -14,6 +14,8 @@
     filter-status-levels="{{ statusAccessLevels|json_encode }}"
     activated-plugins="{{ activatedPlugins|json_encode }}"
     password-strength-validation-rules="{{ passwordStrengthValidationRules|json_encode }}"
+    invite-component="UsersManager.UserInvite"
+    resend-invite-component="UsersManager.ResendInviteModal"
 >
 </div>
