+++ plugins/UsersManager/vue/src/UserEditForm/UserEditForm.less
@@ -1,4 +1,4 @@
-.userEditForm {
+.userEditForm, .userInviteForm {
   .entityList ul li.active a {
     font-weight: bold;
   }
